Updated the create_user_records_in_agents migration to use a transaction making it 100x faster
git-svn-id: https://repository.eol.org/eol/branches/2009.12.11@1577 78829999-583a-0410-bd01-8a9b849fd409

+++ db/migrate/20091217214235_create_user_records_in_agents.rb
@@ -4,15 +4,22 @@ class CreateUserRecordsInAgents < ActiveRecord::Migration
   end
 
   def self.up
-    User.all.each do |user|
-      agent_id = SpeciesSchemaModel.connection.insert(ActiveRecord::Base.eol_escape_sql ["insert into agents (full_name, created_at, updated_at) values (?, now(), now())", user.full_name])
-      user.agent_id = agent_id
-      user.save!
+    SpeciesSchemaModel.connection.begin_db_transaction
+    
+    all_users = User.find_by_sql("SELECT * FROM users")
+    all_users.each do |u|
+      full_name = u.given_name || ""
+      full_name += " " + u.family_name unless u.family_name.blank?
+      
+      agent_id = SpeciesSchemaModel.connection.insert(ActiveRecord::Base.eol_escape_sql ["insert into agents (full_name, created_at, updated_at) values (?, now(), now())", full_name])
+      SpeciesSchemaModel.connection.execute("UPDATE #{User.full_table_name} SET agent_id=#{agent_id} WHERE id=#{u.id}")
     end
+    
+    SpeciesSchemaModel.connection.commit_db_transaction
   end
 
   def self.down
     agent_ids = User.all.map {|u| u.agent_id}
-    execute "delete from agents where id in (#{agent_ids.join(',')})"
+    execute "delete from agents where id in (#{agent_ids.join(',')})" if !agent_ids.empty?
   end
 end
